requirements.txt
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
scipy
streamlit
plotly
groq
python-dotenv
